% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_query.R
\name{send_query}
\alias{send_query}
\title{Send a query string to a SPARQL endpoint}
\usage{
send_query(endpoint_url, query_string, format = "df")
}
\arguments{
\item{endpoint_url}{a string representing URL of SPARQL endpoint}

\item{query_string}{a string of SPARQL codes}

\item{format}{a string that specify return format, default="df"}
}
\value{
A data frame or json data of a query result
}
\description{
Send a query string to a SPARQL endpoint
}
\examples{
endpoint_url <- "https://query.wikidata.org/sparql"
query_string <-'
SELECT ?spaceProbeLabel ?date ?picture
WHERE
{
  ?spaceProbe wdt:P31 wd:Q26529;
  wdt:P18 ?picture;
  wdt:P619 ?date . #mandatory
  SERVICE wikibase:label {
    bd:serviceParam wikibase:language "fr,en" .
  }
}
ORDER BY ?date
LIMIT 10
'

}
