% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_query_from_file.R
\name{send_query_from_file}
\alias{send_query_from_file}
\title{Read a query string from a file and send it to a SPARQL endpoint}
\usage{
send_query_from_file(endpoint_url, file, format = "df")
}
\arguments{
\item{endpoint_url}{a string representing URL of SPARQL endpoint}

\item{file}{a path to a file of SPARQL codes. i.e. ./example.sparql}

\item{format}{a string that specify return format, default="df"}
}
\value{
A data frame or json data of a query result
}
\description{
Read a query string from a file and send it to a SPARQL endpoint
}
\examples{
sparqlfile <- "./examplequery.sparql"
endpoint_url <- "https://query.wikidata.org/sparql"
df <- send_query_from_file(endpoint_url, sparqlfile, format="df")

}
