% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_query.R
\name{show_query}
\alias{show_query}
\title{Read a query string from a file and show it}
\usage{
show_query(file)
}
\arguments{
\item{file}{a path to a file of SPARQL codes. i.e. ./example.sparql}
}
\description{
Read a query string from a file and show it
}
